Add initial data for 'omswarup' and update filter logic
Added a new JSON file `Omswarup.json` with initial properties including `name`, `country_of_origin`, `gender`, `samples`, `accuracy`, and `duration`.

Modified the constructor of the `OutlookHelper6` class in `OutlookHelper6.cs` to accept a boolean parameter `checkRunOutlook` with a default value of `false`. Updated the logic to check if the `OUTLOOK` process is running only if `checkRunOutlook` is `true`.

Updated filter logic in `Page01VM.cs` and `Page02VM.cs` to include a check for `Fname` containing the `SearchText`.

In `Page03VM.cs`, within the `InitAsync` method, changed the call to `AltProcessRunner.RunAsync` to use `ConfigureAwait(false)` and added a delay of 2.5 seconds. Also, within the `WrapAsync` method, added calls to `Synth.SpeakAsync` and `Synth.SpeakFAF` to notify the user about Outlook's status.